Time: 7 ms (82.69%) | Memory: 8.7 MB (99.31%) - LeetSync

+++ 3-longest-substring-without-repeating-characters/longest-substring-without-repeating-characters.cpp
@@ -6,7 +6,7 @@ public:
         string holder;
 
         for (int i=0;i<s.size();i++){
-            int pos = holder.find(s[i]);
+            size_t pos = holder.find(s[i]);
             if (pos == -1) {
                 holder += s[i];
                 longest = max(longest, static_cast<int>(holder.size()));
